+++ cmem_plugin_kafka/utils.py
@@ -4,7 +4,8 @@ import re
 from typing import Dict, Any, Iterator
 from xml.sax.handler import ContentHandler  # nosec B406
 from xml.sax.saxutils import escape  # nosec B406
-
+from urllib.parse import urlparse
+from cmem.cmempy.config import get_cmem_base_uri
 from cmem.cmempy.workspace.projects.resources.resource import get_resource_response
 from cmem.cmempy.workspace.tasks import get_task
 from cmem_plugin_base.dataintegration.context import (
@@ -20,7 +21,7 @@ from cmem_plugin_base.dataintegration.utils import (
 from confluent_kafka import Producer, Consumer, KafkaException
 from confluent_kafka.admin import AdminClient, TopicMetadata, ClusterMetadata
 
-from .constants import KAFKA_TIMEOUT
+from cmem_plugin_kafka.constants import KAFKA_TIMEOUT
 
 
 # pylint: disable-msg=too-few-public-methods
@@ -241,6 +242,13 @@ class KafkaMessageHandler(ContentHandler):
                 operation_desc="messages sent",
             )
         )
+
+
+def get_default_client_id(project_id: str, task_id: str):
+    """return dns:projectId:taskId when client id is empty"""
+    base_url = get_cmem_base_uri()
+    dns = urlparse(base_url).netloc
+    return f"{dns}:{project_id}:{task_id}"
 
 
 def validate_kafka_config(config: Dict[str, Any], topic: str, log: PluginLogger):