+++ tests/test_perf_200K_messages.py
@@ -33,7 +33,7 @@ CONSUMER_DATASET_ID = f"{CONSUMER_DATASET_NAME}"
 KAFKA_CONFIG = get_kafka_config()
 DEFAULT_GROUP = "workflow"
 DEFAULT_TOPIC = "eccenca_kafka_workflow"
-DEFAULT_RESET = "latest"
+DEFAULT_RESET = "earliest"
 RESOURCE_LINK = "https://download.eccenca.com/cmem-plugin-kafka/286K_Message.zip"
 
 
@@ -85,6 +85,7 @@ def test_performance_execution_kafka_producer_consumer(perf_project):
         sasl_username=KAFKA_CONFIG["sasl_username"],
         sasl_password=KAFKA_CONFIG["sasl_password"],
         kafka_topic=DEFAULT_TOPIC,
+        client_id="",
     ).execute([], TestExecutionContext(project_id=PROJECT_NAME))
 
     # Consumer
@@ -97,7 +98,7 @@ def test_performance_execution_kafka_producer_consumer(perf_project):
         sasl_password=KAFKA_CONFIG["sasl_password"],
         kafka_topic=DEFAULT_TOPIC,
         group_id=DEFAULT_GROUP,
-        auto_offset_reset="earliest",
+        auto_offset_reset=DEFAULT_RESET,
     ).execute([], TestExecutionContext(project_id=PROJECT_NAME))
 
     # Ensure producer and consumer are working properly